Added a condition to remove the last 5 minutes of dive samples from a Uwatec Galileo Sol
The Uwatec Galileo Sol records 5 minutes of samples when on the surface
before it considers the dive complete. When we download the samples from
the computer we get those 5 minutes of samples. Thi gives us an
inconsistency between what subsurface and the computer report as the dive
times, etc. This fix will remove the last 5 minutes of samples to make
subsurface and the computer match.

Signed-off-by: Terrance Stanfield <t@hollowcranium.com>